Migration feature now are splitted in 3 files to better organization
